Add ./ctrl: the control plane's record keeper, split from the client
src/ctrl/index.ts is the same shape as the client: a switch over test,
test-results, and session, each action in its own file with its own spec and arg
type, and parseCtrlArgs as the first line. parseCtrlArgs runs Effect's CLI on the
action's flags plus --server-url (SERVER_URL fallback, http/https, no default),
then reads DATABASE_URL and the action's own env through Config. The Linear API
moves to src/linear.ts. connectDatabase and prompt take the url and api key they
used to read from the environment.

client.md covers the guest only; ctrl.md is the full control spec and
ctrl-linear.md the two commands a driving agent needs, rendered into every
Linear issue with the action-first, kebab-case invocation.
