+++ AGENTS.md
@@ -1,4 +1,4 @@
 Two kinds of agent work here.
 
 - A developing agent changes the code. Read [development.md](development.md) and the [field guide](field-guide/index.md).
-- A driving agent uses `./client` to complete a task given in Linear. It never reads or changes code. Read [client.md](client.md); its first 24 lines are the table of contents, consult that first.
+- A driving agent uses `./client` to drive a guest and `./ctrl` to record the result for a task given in Linear. It never reads or changes code. Read [client.md](client.md) and [ctrl.md](ctrl.md); their first lines are a table of contents, consult that first.
