+++ ctrl
@@ -1,0 +1,2 @@
+#!/bin/sh
+exec node --experimental-strip-types "$(dirname "$0")/src/ctrl/index.ts" "$@"
